ExcelMind AI — Chat with Your Customer Data/requirements.txt
streamlit
pandas
numpy
google-generativeai
python-dotenv
openpyxl
